Add lecture content service

+++ services/content/src/main/java/com/genedu/content/repository/SubjectRepository.java
@@ -1,13 +1,12 @@
 package com.genedu.content.repository;
 
-import com.genedu.content.dto.subject.SubjectResponseDTO;
 import com.genedu.content.model.Subject;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
+public interface SubjectRepository extends JpaRepository<Subject, Integer> {
     boolean existsByName(String name);
-    boolean existsByNameAndIdNot(String name, Long id);
+    boolean existsByNameAndIdNot(String name, Integer id);
     List<Subject> findBySchoolClass_Id(Integer schoolClassId);
 }
